Short names:  Support duplicate #include.
Previously, when a header was included a second time with CSNIP_SHORT_NAMES
defined, but CSNIP_SHORT_NAMES was not defined upon the first include, then
short names would not be emitted.  That could be confusing, e.g., when a
csnip header was used in another header.  Change it so that
CSNIP_SHORT_NAMES works even for second inclusion.