build:  Disable BUILD_DOCS by default.
Building with documentation adds the dependency on Doxygen.  It also
creates a slew of warnings currently, due to incomplete documentation.
This can hide other problems, and is also distracting.  Once there are less
warnings, we can consider if we want to reenable this by default.

While here, I notice that the BUILD_DOCS option was unintentionally added
twice.