x:  Add csnip_x_asprintf().
Right now, this function is only provided if there is a system asprintf();
e.g. we don't yet provide an implementation based on snprintf().  So the
advantage over the system asprintf() is only the more uniform API, and the
fact that feature macros such as _GNU_SOURCE need not be defined to use
csnip_x_asprintf().