coro_pth:  Implement.
Implement coro_pth, an implementation of coroutines based on pthreads.  The
preexisting implemenation based on ucontext_t was renamed from csnip_* to
csnip_uctx_*.  A suitable implementation is selected as the default one
(csnip_coro_* namespace) at compile time.  The ucontext_t one is preferred,
if available, since it is faster.

This restores macOS support, which did not work any longer because the
getcontext() API is no longer operational on macOS.