coro_test1:  Another coroutine test.
In some special cases, gcc requires that the stack be executable.  This can
break coroutine implementations which don't take this into account.  This
test covers that case.  Currently the test case fails, since in our
implementation, stacks are not executable.  A fix however is forthcoming.

Test case can only be used with gcc, and not clang, since it relies on
gcc's nested functions feature.

While here, rename coro_test -> coro_test0;  this is more in line with the
other tests.