clopts:  Support for freeing memory after use.
This change implements csnip_clopts_clear(), as well as adds the do_clear
flag to csnip_clopts_process().  Previously, the small amount of memory
allocated when adding options to clopts was never freed.  Since that
memory is small and constant, it was considered not worth freeing it.  But
on the other hand; memory debuggers will flag it as a leak, and sometimes
the goal is to be entirely leak-free.  This change makes this possible.

Unfortunately, previous users of csnip_clopts_process() need to be updated
to the new signature.