clopts:  Require C11 only for generic macros.
Previously, the compiled parts of clopts would require C11, since they made
use of the csnip/limits library.  That resulted in all of csnip requiring
C11, which is a problem.  Avoid relying on csnip for compiled parts of
clopts.

C11 is still required for using macros such as clopts_AddVar() which are
based on _Generic; that is ok, however:  C11 is unavoidable in this case.
If C11 is not acceptable, then clopts can still be used, but not with the
convenience of the type generic macros.