limits:  Work around broken compilers.
A number of compilers get it wrong that if, say "int" and "long int" have
the same representation in a particular implementation, they are still
distinct types, thus _Generic must distinguish the two.

To avoid this problem, instead of

	_Generic(x, type1: foo, type2: bar),

we do something like

	_Generic(x, type1: foo, default: _Generic(x, type2: bar)),

where the previous problem does not occur.