meanvar:  Fix empty macro argument warning.
A macro foo(x) taking one argument can be invoked as foo(), meaning that
its argument is the empty string.  Some compiler warn about this, as it
may be the case that the user has forgotten to supply an argument.

We work around this warning by making the macro accept two arguments (i.e.,
foo(x,y)), the first of which is a dummy.   This works because compilers
don't warn about foo(dummy,).