all:  Build fixes for MSVC.
This is a very old set of changes, implementing better MSVC support.  At
this point I don't know if it's enough; probably there are still more
changes needed to get things to work on MSVC, but it's a big step in the
right direction, and it's better to commit this rather than keep it
lying around forever in my working copy.