ringbuf2:  Add *_areas functions.
Sometimes, it is useful to get all the memory areas readable or writable
into a ringbuffer.  This can be used e.g. to use pwrite() or to inspect
all the ringbuffer entries without consuming them, or to obtain the
whole contents at once, e.g. because one must hold onto a mutex to
obtain them.