all:  Make multithreading support configurable.
This is useful in cases where multi threading is not used.  It is also
currently necessary to turn off threading when compiling with MSVC as
that compiler does not have pthreads.

While here, remove the broken CSNIP_CONF__EMULATE_{READ,WRITE}V
macros from the configuration file; they're not working.