arr:  Rename Free -> Deinit.
While this function frees the underlying array, and resets the entries,
the a, n, and cap variables themselves are not freed (not that this
necessarily makes sense).

Using arr alone this distinction is not that important, but a user of
arr might want to define a _free function that actually frees the entire
container, including the meta data.

The naming is also more consistent;  deinit() is the complement to
init().